requirements.txt
Flask>=2.0
requests>=2.28
icalendar>=4.0
python-dateutil>=2.8
pytz
gunicorn